+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/repository/RdbmsDeploymentIdRepository.java
@@ -22,7 +22,6 @@ import java.sql.Types;
 
 import javax.sql.DataSource;
 
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.util.Assert;
 
@@ -43,25 +42,19 @@ public class RdbmsDeploymentIdRepository extends AbstractRdbmsKeyValueRepository
 	}
 
 	@Override
-	public String findOne(String deploymentKey) {
-		Assert.hasText(deploymentKey, "deploymentKey must not be empty nor null");
-		try {
-			return jdbcTemplate.queryForObject(findAllWhereClauseByKey, rowMapper, deploymentKey);
-		}
-		catch (EmptyResultDataAccessException e) {
-			return null;
-		}
-	}
-
-	@Override
-	public void save(String deploymentKey, String deploymentId) {
-		Object[] insertParameters = new Object[]{deploymentKey, deploymentId};
+	public void save(String key, String deploymentId) {
+		Object[] insertParameters = new Object[]{key, deploymentId};
 		jdbcTemplate.update(saveRow, insertParameters, new int[]{Types.VARCHAR, Types.VARCHAR});
 	}
 
 	@Override
-	public void delete(String deploymentKey) {
-		Assert.hasText(deploymentKey, "deploymentKey must not be empty nor null");
-		jdbcTemplate.update(deleteFromTableByKey, deploymentKey);
+	public String save(String key) {
+		throw new UnsupportedOperationException();
+	}
+
+	@Override
+	public void delete(String name) {
+		Assert.hasText(name, "name must not be empty nor null");
+		jdbcTemplate.update(deleteFromTableByKey, name);
 	}
 }
