+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/repository/RdbmsStreamDefinitionRepository.java
@@ -31,7 +31,7 @@ import org.springframework.util.Assert;
  *
  * @author Ilayaperumal Gopinathan
  */
-public class RdbmsStreamDefinitionRepository extends AbstractRdbmsKeyValuePagingRepository<StreamDefinition> implements StreamDefinitionRepository {
+public class RdbmsStreamDefinitionRepository extends AbstractRdbmsKeyValueRepository<StreamDefinition> implements StreamDefinitionRepository {
 
 	public RdbmsStreamDefinitionRepository(DataSource dataSource) {
 		super(dataSource, "STREAM_", "DEFINITIONS", new RowMapper<StreamDefinition>() {
